2.4 working starting 2.5

+++ oblig5/src/main/java/no/hiof/erikvs/model/NaturalSatellite.java
@@ -75,7 +75,7 @@ public abstract class NaturalSatellite extends CelestialBody {
         return orbitalPeriod;
     }
 
-    public void setOrbitalPeriod(int orbitalPeriod) {
+    public void setOrbitalPeriod(double orbitalPeriod) {
         this.orbitalPeriod = orbitalPeriod;
     }
